+++ cgworkshop/src/MaxFlow/graph.h
@@ -66,7 +66,7 @@ public:
 
 	/* Type of edge weights.
 	   Can be changed to char, int, float, double, ... */
-	typedef short captype;
+	typedef double captype;
 	/* Type of total flow */
 	typedef int flowtype;
